Beim löschen eines users wird überprüft, ob der löscher ein admin ist

+++ Project Herakles/Project Herakles/MainForm.cs
@@ -27,13 +27,20 @@ namespace Project_Herakles
 
         private void deleteKundeButton_Click(object sender, EventArgs e)
         {
-
+            
             LoginForm loginForm = new LoginForm();
             loginForm.ShowDialog();
             if (loginForm.login == true)
-            {
+            { 
                 DBController dBController = new DBController();
-                dBController.deleteKunde(loginForm.loginName);
+                if (dBController.checkAdmin(loginForm.loginName) == true)
+                {
+                    dBController.deleteKunde(loginForm.loginName);
+                }
+                else
+                {
+                    MessageBox.Show("Sie haben keine Rechte!");
+                }
             }
             else
             {
